% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\name{calculate_en_score}
\alias{calculate_en_score}
\title{Calculate En-score (Error normalized)}
\usage{
calculate_en_score(x, x_pt, U_x, U_xpt)
}
\arguments{
\item{x}{Participant result.}

\item{x_pt}{Assigned value.}

\item{U_x}{Expanded uncertainty of participant's result.}

\item{U_xpt}{Expanded uncertainty of the assigned value.}
}
\value{
En-score value.
}
\description{
En = (x - x_pt) / sqrt(U_x^2 + U_xpt^2)
}
\details{
Reference: ISO 13528:2022, Section 10.5
}
\examples{
# En-score uses expanded uncertainties (k=2)
en <- calculate_en_score(x = 10.5, x_pt = 10.0, U_x = 0.4, U_xpt = 0.2)
cat("En-score:", en, "Eval:", evaluate_en_score(en))

}
\seealso{
\code{\link{evaluate_en_score}}, \code{\link{classify_with_en}}
}
