% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_homogeneity.R
\name{calculate_stability_criterion_expanded}
\alias{calculate_stability_criterion_expanded}
\title{Calculate expanded stability criterion}
\usage{
calculate_stability_criterion_expanded(c_criterion, u_hom_mean, u_stab_mean)
}
\arguments{
\item{c_criterion}{Base stability criterion}

\item{u_hom_mean}{Uncertainty of homogeneity mean}

\item{u_stab_mean}{Uncertainty of stability mean}
}
\value{
The expanded stability criterion
}
\description{
c_stab_expanded = c_criterion + 2 * sqrt(u_hom_mean^2 + u_stab_mean^2)
}
