% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_homogeneity.R
\name{calculate_u_hom}
\alias{calculate_u_hom}
\title{Calculate uncertainty contribution from homogeneity}
\usage{
calculate_u_hom(ss)
}
\arguments{
\item{ss}{Between-sample standard deviation from homogeneity study}
}
\value{
Uncertainty contribution from homogeneity
}
\description{
u_hom = ss (between-sample standard deviation)
}
\details{
Reference: ISO 13528:2022, Section 9.5
}
