% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_homogeneity.R
\name{calculate_u_stab}
\alias{calculate_u_stab}
\title{Calculate uncertainty contribution from stability}
\usage{
calculate_u_stab(diff_hom_stab, c_criterion)
}
\arguments{
\item{diff_hom_stab}{Absolute difference between stability and homogeneity means}

\item{c_criterion}{Stability criterion}
}
\value{
Uncertainty contribution from stability
}
\description{
u_stab = diff_hom_stab / sqrt(3) (if criterion not met)
or 0 (if criterion is met)
}
\details{
Reference: ISO 13528:2022, Section 9.5
}
