% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\name{calculate_z_prime_score}
\alias{calculate_z_prime_score}
\title{Calculate z'-score (z-prime score)}
\usage{
calculate_z_prime_score(x, x_pt, sigma_pt, u_xpt)
}
\arguments{
\item{x}{Participant result.}

\item{x_pt}{Assigned value.}

\item{sigma_pt}{Standard deviation for proficiency assessment.}

\item{u_xpt}{Standard uncertainty of the assigned value.}
}
\value{
z'-score value.
}
\description{
z' = (x - x_pt) / sqrt(sigma_pt^2 + u_xpt^2)
}
\details{
Reference: ISO 13528:2022, Section 10.3
}
\examples{
# z'-score accounts for uncertainty in assigned value
zprime <- calculate_z_prime_score(x = 10.5, x_pt = 10.0, sigma_pt = 0.5, u_xpt = 0.1)

}
\seealso{
\code{\link{calculate_z_score}}, \code{\link{calculate_zeta_score}}
}
