% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\name{calculate_z_score}
\alias{calculate_z_score}
\title{Calculate z-score}
\usage{
calculate_z_score(x, x_pt, sigma_pt)
}
\arguments{
\item{x}{Participant result.}

\item{x_pt}{Assigned value.}

\item{sigma_pt}{Standard deviation for proficiency assessment.}
}
\value{
z-score value.
}
\description{
z = (x - x_pt) / sigma_pt
}
\details{
Reference: ISO 13528:2022, Section 10.2
}
\examples{
# Calculate z-score for a participant
z <- calculate_z_score(x = 10.5, x_pt = 10.0, sigma_pt = 0.5)
cat("z-score:", z)  # 1.0 (Satisfactorio)

}
\seealso{
\code{\link{calculate_z_prime_score}}, \code{\link{evaluate_z_score}}
}
