% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\name{calculate_zeta_score}
\alias{calculate_zeta_score}
\title{Calculate zeta-score}
\usage{
calculate_zeta_score(x, x_pt, u_x, u_xpt)
}
\arguments{
\item{x}{Participant result.}

\item{x_pt}{Assigned value.}

\item{u_x}{Standard uncertainty of participant's result.}

\item{u_xpt}{Standard uncertainty of the assigned value.}
}
\value{
zeta-score value.
}
\description{
zeta = (x - x_pt) / sqrt(u_x^2 + u_xpt^2)
}
\details{
Reference: ISO 13528:2022, Section 10.4
}
\examples{
# zeta-score uses participant's uncertainty
zeta <- calculate_zeta_score(x = 10.5, x_pt = 10.0, u_x = 0.2, u_xpt = 0.1)

}
\seealso{
\code{\link{calculate_en_score}}
}
