% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\name{evaluate_en_score}
\alias{evaluate_en_score}
\title{Evaluate En-score}
\usage{
evaluate_en_score(en)
}
\arguments{
\item{en}{En-score value}
}
\value{
Character string with evaluation category
}
\description{
Classifies En-score performance:
\itemize{
\item |En| <= 1: Satisfactorio (Satisfactory)
\item |En| > 1: No satisfactorio (Unsatisfactory)
}
}
