% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\name{evaluate_en_score_vec}
\alias{evaluate_en_score_vec}
\title{Vectorized En-score evaluation}
\usage{
evaluate_en_score_vec(en)
}
\arguments{
\item{en}{Vector of En-score values}
}
\value{
Character vector with evaluation categories
}
\description{
Vectorized En-score evaluation
}
