% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\name{evaluate_z_score_vec}
\alias{evaluate_z_score_vec}
\title{Vectorized z-score evaluation}
\usage{
evaluate_z_score_vec(z)
}
\arguments{
\item{z}{Vector of score values}
}
\value{
Character vector with evaluation categories
}
\description{
Vectorized z-score evaluation
}
