% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\docType{data}
\name{PT_EN_CLASS_COLORS}
\alias{PT_EN_CLASS_COLORS}
\title{Classification colors for combined z/En evaluation}
\format{
An object of class \code{character} of length 9.
}
\usage{
PT_EN_CLASS_COLORS
}
\description{
Classification colors for combined z/En evaluation
}
\keyword{datasets}
