% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_scores.R
\docType{data}
\name{PT_EN_CLASS_LABELS}
\alias{PT_EN_CLASS_LABELS}
\title{Classification labels for combined z/En evaluation}
\format{
An object of class \code{character} of length 7.
}
\usage{
PT_EN_CLASS_LABELS
}
\description{
Classification labels for combined z/En evaluation
}
\keyword{datasets}
