% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptcalc-package.R
\docType{package}
\name{ptcalc-package}
\alias{ptcalc-package}
\alias{ptcalc}
\title{ptcalc: Proficiency Testing Calculations}
\description{
Functions for proficiency testing analysis per ISO 13528:2022 and ISO 17043:2023.
}
\section{Robust Statistics}{

\itemize{
\item \code{\link{calculate_niqr}} - Normalized IQR
\item \code{\link{calculate_mad_e}} - Scaled MAD
\item \code{\link{run_algorithm_a}} - Algorithm A for robust mean/SD
}
}

\section{Score Calculations}{

\itemize{
\item \code{\link{calculate_z_score}} - z-score
\item \code{\link{calculate_z_prime_score}} - z'-score
\item \code{\link{calculate_zeta_score}} - zeta-score
\item \code{\link{calculate_en_score}} - En-score
\item \code{\link{evaluate_z_score}} - Evaluate z-type scores
\item \code{\link{evaluate_en_score}} - Evaluate En-scores
}
}

\section{Homogeneity/Stability}{

\itemize{
\item \code{\link{calculate_homogeneity_stats}} - Homogeneity statistics
\item \code{\link{calculate_stability_stats}} - Stability statistics
\item \code{\link{evaluate_homogeneity}} - Evaluate homogeneity criterion
\item \code{\link{evaluate_stability}} - Evaluate stability criterion
}
}

\author{
\strong{Maintainer}: Wilson Rafael Salas Chavez \email{wrsalasc@unal.edu.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID}) (Lead Developer)

Authors:
\itemize{
  \item Wilson Rafael Salas Chavez \email{wrsalasc@unal.edu.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID}) (Lead Developer)
}

Other contributors:
\itemize{
  \item Universidad Nacional de Colombia (Laboratorio CALAIRE - Titular de Derechos) [copyright holder, funder]
  \item Instituto Nacional de Metrología (Entidad Aliada) [funder]
}

}
