% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_homogeneity.R
\name{calculate_homogeneity_criterion_expanded}
\alias{calculate_homogeneity_criterion_expanded}
\title{Calculate expanded homogeneity criterion}
\usage{
calculate_homogeneity_criterion_expanded(sigma_pt, u_sigma_pt)
}
\arguments{
\item{sigma_pt}{Standard deviation for proficiency assessment (from MADe)}

\item{u_sigma_pt}{Uncertainty of sigma_pt}
}
\value{
The expanded criterion value
}
\description{
c_expanded = c_criterion * sqrt(1 + (u_sigma_pt/sigma_pt)^2)
}
\details{
Reference: ISO 13528:2022, Section 9.2.4
}
